Reduce warning in map class

+++ Super-Mario-DX10/Map.cpp
@@ -41,7 +41,7 @@ void CMap::Render()
 		{
 			int index = TileMap[CurrentRow][CurrentColumn] - 1;
 			if (index < TotalTiles)
-				Tiles.at(index)->Draw(CurrentColumn * TILE_WIDTH,(CurrentRow * TILE_HEIGHT));
+				Tiles.at(index)->Draw(float(CurrentColumn * TILE_WIDTH),float(CurrentRow * TILE_HEIGHT));
 		}
 }
 void CMap::SetTileMapData(int** TileMapData)
